+++ src/transformers/compare.ts
@@ -5,7 +5,7 @@ import {
   codapValueToString,
   getAttributeDataFromDataset,
 } from "./util";
-import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
+import { DDTransformerState } from "../components/transformer-template/DataDrivenTransformer";
 import { getContextAndDataSet } from "../lib/codapPhone";
 import { readableName } from "../transformer-components/util";
 import { uniqueName } from "../lib/utils/names";
