+++ src/transformers/count.ts
@@ -8,7 +8,7 @@ import {
   validateAttribute,
 } from "./util";
 import { uniqueName } from "../lib/utils/names";
-import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
+import { DDTransformerState } from "../components/transformer-template/DataDrivenTransformer";
 import { getContextAndDataSet } from "../lib/codapPhone";
 import { readableName } from "../transformer-components/util";
