+++ src/transformers/fold.ts
@@ -8,7 +8,7 @@ import {
 } from "./util";
 import { evalExpression, getContextAndDataSet } from "../lib/codapPhone";
 import { uniqueName } from "../lib/utils/names";
-import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
+import { DDTransformerState } from "../components/transformer-template/DataDrivenTransformer";
 import { parenthesizeName, readableName } from "../transformer-components/util";
 
 type FoldFunction = (
