+++ src/transformers/join.ts
@@ -1,6 +1,6 @@
 import { DataSet, TransformationOutput } from "./types";
 import { uniqueName } from "../lib/utils/names";
-import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
+import { DDTransformerState } from "../components/transformer-template/DataDrivenTransformer";
 import { getContextAndDataSet } from "../lib/codapPhone";
 import { readableName } from "../transformer-components/util";
 import {
