+++ src/transformers/transformColumn.ts
@@ -1,6 +1,6 @@
 import { CodapLanguageType, DataSet, TransformationOutput } from "./types";
 import { evalExpression, getContextAndDataSet } from "../lib/codapPhone";
-import { DDTransformerState } from "../transformer-components/DataDrivenTransformer";
+import { DDTransformerState } from "../components/transformer-template/DataDrivenTransformer";
 import { readableName } from "../transformer-components/util";
 import {
   reportTypeErrorsForRecords,
